% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr2vec.R
\name{chr2vec}
\alias{chr2vec}
\title{Helper for mHM_read.nml}
\usage{
chr2vec(chr)
}
\arguments{
\item{chr}{vector of characters to be splited and trimed.}
}
\value{
splited, trimed characters in list.
}
\description{
Helper for mHM_read.nml
}
\examples{
chr2vec("timestep   =   1")
chr2vec("gaugefile   =  \\"basin01.txt\\"")
chr2vec("aggregation   =  .TRUE.")
chr2vec( "PETcorr  =  1,1.2,0,95  ")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
