% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_asc2netCDF.R
\name{mHM_asc2netCDF}
\alias{mHM_asc2netCDF}
\title{convert ascii to netCDF.}
\usage{
mHM_asc2netCDF(rst = NA, outfile = "out.nc", crs, varname, varunit,
  longname, xname = "lon", yname = "lat", project = FALSE)
}
\arguments{
\item{rst}{raster input object, either raster obj or files formate for read in with \code{\link[raster]{raster}}.}

\item{outfile}{character, output file name, default: "out.nc".}

\item{crs}{either character proj4 coordinate projection definition, e.g. "+proj=utm +zone=30 +ellps=intl +units=m +no_defs" for "epsg:23030". Or character epsg coordinate projection definition, e.g. "epsg:23030".}

\item{varname}{character, variable name to be set in netCDF.}

\item{varunit}{character, variable unit to be set in netCDF.}

\item{longname}{character, variable long name to be set in netCDF.}

\item{xname}{character, x coordinate name, default: "lon"}

\item{project}{boolean, project data, DEFAULT: FALSE}

\item{xname}{character, y coordinate name, default: "lat"}
}
\value{
netCDF written to outfile.
}
\description{
\code{mHM_asc2netCDF} is converting ascii spatial file format to netCDF.
}
\examples{

}
\references{

}
\seealso{
\code{\link[raster]{raster}}, \code{\link[raster]{crs}}, \code{\link[raster]{writeRaster}}, \code{\link[raster]{projectRaster}}
}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
