% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_avg2zoo.R
\name{mHM_avg2zoo}
\alias{mHM_avg2zoo}
\title{get summarized zoo object for mHM basin.}
\usage{
mHM_avg2zoo(con, var, datetime, fun)
}
\arguments{
\item{con}{connection to netCDF.}

\item{var}{variable name.}

\item{datetime}{Date, POSIX or Yearmon object.}

\item{fun}{function to summarize/aggregate basin.}
}
\value{
Zoo object.
}
\description{
\code{mHM_avg2zoo} is retrieving a basin summary, depicted by function \code{fun}, and giving back a zoo object.
}
\examples{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{
\code{\link{mHM_getDateTime}}
}
\keyword{}

