% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_avg2zoo.R
\name{mHM_avgnetcdf2zoo}
\alias{mHM_avgnetcdf2zoo}
\title{get summarized zoo object for mHM basin.}
\usage{
mHM_avgnetcdf2zoo(ncfile, var, fun, out_unit = "d")
}
\arguments{
\item{ncfile}{mHM netCDF output file.}

\item{var}{variable name.}

\item{fun}{function to summarize/aggregate basin.}

\item{out_unit}{definition of datetime unit out; 'h'=hour ,'d'=day, 'm'=month.}
}
\value{
Zoo object.
}
\description{
\code{mHM_avg2zoo} is retrieving a basin summary, depicted by function \code{fun}, and giving back a zoo object.
}
\examples{

}
\references{

}
\seealso{
\code{\link{mHM_getDateTime}}
}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
