% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_checkT.R
\name{mHM_checkT}
\alias{mHM_checkT}
\title{Check minimum and maximum air temperature for consistency in NetCDF.}
\usage{
mHM_checkT(nc.tmin, nc.tmax)
}
\arguments{
\item{nc.tmin}{netCDF of minimum air temperature.}

\item{nc.tmax}{netCDF of maximum air temperature.}
}
\value{
re-write input netCDF files with consistent air temperatures.
}
\description{
\code{mHM_checkT} proves if minimum air temperature is lower than maximum air temperature. 
If this is not the case values are interchanged. Input format is netCDF.
}
\details{
Runing mHM the first time for some netCDF products (e.g. Spain02, E-OBS) the warning Tmax smaller than Tmin (or wise versa) will be printed. 
This function makes consistant minmum and maximum air temperature fields in netCDF format.
}
\examples{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

