% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_enlargeNETCDF.R
\name{mHM_enlargeNETCDF}
\alias{mHM_enlargeNETCDF}
\title{FUNCTION_TITLE}
\usage{
mHM_enlargeNETCDF(netcdf, x, y, var, fixcorner = 1)
}
\arguments{
\item{netcdf}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{fixcorner}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RNetCDF]{open.nc}},\code{\link[RNetCDF]{var.get.nc}},\code{\link[RNetCDF]{create.nc}},\code{\link[RNetCDF]{dim.def.nc}},\code{\link[RNetCDF]{var.def.nc}},\code{\link[RNetCDF]{att.copy.nc}},\code{\link[RNetCDF]{var.put.nc}},\code{\link[RNetCDF]{close.nc}}
}
