% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_enlcutgrid.R
\name{mHM_enlcutgrid}
\alias{mHM_enlcutgrid}
\title{enlarge or cut grid}
\usage{
mHM_enlcutgrid(wd, ncol, nrow)
}
\arguments{
\item{wd}{working directory}

\item{ncol}{integer, requested number of columns}

\item{nrow}{integer, requested number of rows}
}
\value{
writting new ,asc file in mHM format
}
\description{
\code{mHM_enlcutgrid} is enlarging or cutting grids (.asc) according to given column and row number in a working dirctoy to be specified
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[raster]{raster}},\code{\link[raster]{getValues}},\code{\link[raster]{setValues}}
 \code{\link[mHMr]{mHM_formatASC}}
}
