% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_extrapolMat.R
\name{mHM_extrapolMat}
\alias{mHM_extrapolMat}
\title{Simple extrapolation arrays in invalid locations.}
\usage{
mHM_extrapolMat(data, ext)
}
\arguments{
\item{data}{matrix.}

\item{ext}{radius within grid cells are extended, starting from valid values.}
}
\value{
matrix.
}
\description{
\code{mHM_extrapolMat} extrapolates matrix values in invalid location, e.g. originating from 
meteorological input fields for mHM (netCDF format).
}
\details{
Runing mHM the first time for some netCDF products (e.g. Spain02, E-OBS) the warning
no meteorological input available for whole simulation domain will be printed. 
This function extrapolates grids in invalid locations within a radius given by \code{ext}.
Extrapolation is performed by a mean value of the surrounding cells of the invalid location.
}
\examples{
# read netcdf file minimum air temperature
pkg_path <- path.package("mHMr")
tmin_netcdf <- file.path(pkg_path, "data", "tmin.nc")
tmin_op <-  RNetCDF::open.nc(tmin_netcdf)
tmin <- RNetCDF::var.get.nc(tmin_op,"tmin")
image(tmin[,,1])

tmin1 <- mHM_extrapolMat(tmin[,,1], ext=1)
image(tmin1,  zlim = c(2.5,4.5))

tmin2 <- mHM_extrapolMat(tmin[,,1], ext=2)
image(tmin2,  zlim = c(2.5,4.5))



}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

