% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_formatASC.R
\name{mHM_formatASC}
\alias{mHM_formatASC}
\title{Reformat ASCII file.}
\usage{
mHM_formatASC(inASC)
}
\arguments{
\item{inASC}{character vector, gives location of raster files in ASCII format.}
}
\description{
\code{mHM_formatASC} is rewriting the header of an ASCII file being written by \code{\link{raster::writeRaster}}.
}
\details{
ASCII files written by \code{\link{raster::writeRaster}} do not fullfil conventions for mHM input ASCII files.
This function rewrites the header and find&replaces additional zeros appended to -9999 values.
}
\examples{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{ascii}

