% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_getnetCDFvarnames.R
\name{mHM_getnetCDFvarnames}
\alias{mHM_getnetCDFvarnames}
\title{Retrieve netCDF variable names.}
\usage{
mHM_getnetCDFvarnames(ncfile)
}
\arguments{
\item{ncfile}{netCDF file name plus path.}
}
\description{
\code{mHM_getnetCDFvarnames} is reading variable names from a netCDF file.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
