% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_GOF4period.R
\name{mHM_GOF4period}
\alias{mHM_GOF4period}
\title{Calculate GOFs for time intervals.}
\usage{
mHM_GOF4period(zoo_obj, start = NA, end = NA, plot = FALSE)
}
\arguments{
\item{zoo_obj}{zoo object containing observed and simulated daily discharge.}

\item{start}{start date, either Date object or character in the following format yyyy-mm-dd.}

\item{end}{end date, either Date object or character in the following format yyyy-mm-dd.}

\item{plot}{boolean, if TRUE ggof plot from hydroGOF package.}
}
\value{
numeric vector of GOF values.
}
\description{
\code{mHM_GOF4period} returns Goodness of Fit (GOF) measures for specific time intervals. 
GOF measures are calculated with \code{hydroGOF::gof}.
}
\details{
Besides calculation of GOFs \code{mHM_GOF4period} provides ploting facilities from \code{hydroGOF::ggof}. 
By default three plots are produced: daily and monthly aggregation, seasonal aggregation, 
all of them comparing simulated and observed discharge time series.
}
\examples{
data(Qout_mHM)
cal_gof <- mHM_GOF4period(zoo_obj = Qout_mHM, start = "2000-01-01", end = "2008-03-30", plot = F)

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

