% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_maskRaster.R
\name{mHM_maskRaster}
\alias{mHM_maskRaster}
\title{Mask raster maps with same projection and extent.}
\usage{
mHM_maskRaster(mask, inRaster, outRaster)
}
\arguments{
\item{mask}{character, locating and defining raster format (either ASCII or GEOtif) or Raster* object.}

\item{inRaster}{character vector, gives location and format of raster files (either ASCII or GEOtif), 
with same projection and extent as provided raster mask.}

\item{outRaster}{character vector, defining raster format and location of the outfile, same length as \code{inRaster}.}
}
\value{
raster in given format (\code{outRaster}) written to outRaster
}
\description{
\code{mHM_maskRaster} masks a raster with same projection and extent as the provided mask.
}
\examples{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

