% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_matchgrid.R
\name{mHM_matchgrid}
\alias{mHM_matchgrid}
\title{match grids}
\usage{
mHM_matchgrid(grids, mask_grid, mask_var = "pre",
  proj4 = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs")
}
\arguments{
\item{grids}{grid files to be masked}

\item{mask_grid}{grid mask}

\item{mask_var}{grid mask variable, Default: 'pre'}

\item{proj4}{PARAM_DESCRIPTION, Default: '+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +units=m +no_defs'}
}
\value{
grids will be written back to original file
}
\description{
\code{mHM_matchgrid} does mask grids, e.g. morphological maps (.asc) by netCDF meteorological mHM input.
}
\details{
grids will be backed up before processing
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[raster]{raster}},\code{\link[raster]{crs}},\code{\link[raster]{resample}},\code{\link[raster]{mask}},\code{\link[raster]{writeRaster}}
}
