% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_plotTSaggr.R
\name{mhm_plotTSaggr}
\alias{mhm_plotTSaggr}
\title{plot comparing obs and sim for single var.}
\usage{

  mhm_plotTSaggr(obs = "/Users/brennerj/tmp/eve_f2_data/data/processed/Jucar/EC_sites/06/input/optional_data/level11/et.nc",
  sim = c("/Users/brennerj/tmp/eve_f2_work/jucar_eddy_out/06/site_meteo/01/01/output/mHM_Fluxes_States.nc",
  "/Users/brennerj/tmp/eve_f2_home/mHMruns/002_Jucar_EC/06/001/out/mHM_Fluxes_States.nc"),
  obs_var = "et", sim_var = "aET", aggr = "month", aggr_fun = sum,
  outfile = "out.pdf", simnames = c("optim_Q", "optim_ET"),
  cols = c(rgb(1, 0, 0, 0.7), rgb(0, 1, 0, 0.7), rgb(0, 0, 1, 0.7)),
  cumsums = F)
}
\arguments{
\item{obs}{.}

\item{sim}{.}

\item{obs_var}{.}

\item{sim_var}{.}

\item{aggr}{.}

\item{aggr_fun}{.}

\item{outfile}{.}

\item{simnames}{.}

\item{cols}{.}

\item{cumsums}{.}
}
\value{
data.frame containing GOF measures (KGE, NSE).
}
\description{
\code{mhm_plotTSaggr} creates a plot (in pdf) comparing .
}
\details{
Runing mHM the first time for some netCDF products (e.g. Spain02, E-OBS) the warning
no meteorological input available for whole simulation domain will be printed. 
This function extrapolates grids in invalid locations within a radius given by \code{ext}.
Extrapolation is performed by a mean value of the surrounding cells of the invalid location.
}
\examples{


}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
