% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_prolongRintime.R
\name{mHM_prolongRintime}
\alias{mHM_prolongRintime}
\title{prolonge netCDF climatology}
\usage{
mHM_prolongRintime(r, nc_time, nc_time_name, nc_out, nc_out_name, standard_name,
  long_name, units)
}
\arguments{
\item{r}{raster file to be prolonged (climatology)}

\item{nc_time}{netCDF file, will be copied to nc_out}

\item{nc_time_name}{variable name of netCDF file nc_time}

\item{nc_out}{netCDF file, prolonged data will be set in}

\item{nc_out_name}{variable name of netCDF file nc_out}

\item{standard_name}{standard name in nc_out}

\item{long_name}{long name in nc_out}

\item{units}{units in nc_out}
}
\value{
nc_out
}
\description{
prolongue netCDF (daily) climatology for specific time period given by netCDF file
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[mHMr]{mHM_getDateTime}}
 \code{\link[raster]{values}}
 \code{\link[RNetCDF]{open.nc}},\code{\link[RNetCDF]{var.put.nc}},\code{\link[RNetCDF]{var.rename.nc}},\code{\link[RNetCDF]{att.put.nc}},\code{\link[RNetCDF]{close.nc}}
}
