% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_read.nml.R
\name{mHM_read.nml}
\alias{mHM_read.nml}
\title{Read mHM daily_discharge.out in zoo object.}
\usage{
mHM_read.nml(simpath)
}
\arguments{
\item{simpath}{path to mHM simulation folder, where mhm.nml is located.}
}
\value{
list of mHM namelist entries.
}
\description{
Read mHM daily_discharge.out in zoo object.
}
\examples{
nml <- mHM_read.nml(simpath="your/path")

}

