% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhm_read_ddsresults.R
\name{mhm_read_ddsresults}
\alias{mhm_read_ddsresults}
\title{Read best GOF DDS.}
\usage{
mhm_read_ddsresults(file, NRlines2read = 100)
}
\arguments{
\item{file}{file name with full path.}

\item{NRlines2read}{integer, number of lines to read from end of file, default=100.}
}
\value{
Best GOF value of DDS calibration run
}
\description{
\code{mhm_read_ddsresults} reads best GOF value of DDS calibration run.
}
\examples{
nml <- mhm_read_ddsresults(file="your/path/dds_results.out")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
