% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhm_read_paramset.R
\name{mhm_read_paramset}
\alias{mhm_read_paramset}
\title{Read best parameter set.}
\usage{
mhm_read_paramset(file)
}
\arguments{
\item{file}{file name with full path.}
}
\value{

}
\description{
\code{mhm_read_paramset} reads best parameterset of mHM calibration run.
}
\examples{
nml <- mhm_read_paramset(file="your/path/FinalParam.out")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
