% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhm_read_sceresults.R
\name{mhm_read_sceresults}
\alias{mhm_read_sceresults}
\title{Read best GOF SCE}
\usage{
mhm_read_sceresults(file)
}
\arguments{
\item{file}{file name with full path.}
}
\value{
Best GOF value of SCE calibration run
}
\description{
\code{mhm_read_sceresults} reads best GOF value of SCE calibration run.
}
\examples{
nml <- mhm_read_sceresults(file="your/path/sce_results.out")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
