% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_readNCvar.R
\name{mHM_readNCvar}
\alias{mHM_readNCvar}
\title{read netCDF variable.}
\usage{
mHM_readNCvar(ncfile, var)
}
\arguments{
\item{ncfile}{character, netCDF file path.}

\item{var}{character, variable name.}
}
\value{
array
}
\description{
\code{mHM_readNCvar} is reading time series data of a specific variable from netCDF
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
