% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_readNCvar2zoo.R
\name{mHM_readNCvar2zoo}
\alias{mHM_readNCvar2zoo}
\title{read netCDF variable to zoo.}
\usage{
mHM_readNCvar2zoo(ncfile, var, out_unit = "d")
}
\arguments{
\item{ncfile}{character, netCDF file path.}

\item{var}{character, variable name.}
}
\value{
array
}
\description{
\code{mHM_readNCvar2zoo} is reading time series data of a specific variable from netCDF. Use function only for data set with a single valid grid cell.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
