% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_recalcAlpha.R
\name{mHM_recalcAlpha}
\alias{mHM_recalcAlpha}
\title{Back-calculate $alpha$.}
\usage{
mHM_recalcAlpha(swc, pwp, sat)
}
\arguments{
\item{swc}{array 3d, containing soil moisture data [mm] for soil layer \eqn{i}. Third dimension is time.}

\item{pwp}{array 2d, containing permanent wilting point [mm] for soil layer \eqn{i}.}

\item{sat}{array 2d, containing soil moisture limit (saturation, porosity) [mm] for soil layer \eqn{i}.}
}
\value{
array of $alpha$ values for soil layer \eqn{i}.
}
\description{
\code{mHM_recalcAlpha} is back-calculating the \eqn{\alpha} parameter for one soil layer \eqn{i} of a mHM model simulation. 
Back-calculation is based on soil water content (\eqn{x_{3}}), permanent wilting point (\eqn{\beta_{15}}) and porosity (\eqn{\beta_{16}}).
}
\details{
\eqn{\alpha} is calculated as follows: 
\deqn{\alpha = \frac{x_{3}(t-1) -  \beta_{15}}{\beta_{16} - \beta_{15}}}
}
\examples{

}
\references{

}
\seealso{
\code{\link{mHM_vargetmulti}}
}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
