% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_reclassint.R
\name{mHM_reclassint}
\alias{mHM_reclassint}
\title{Reclassify raster with lookuptable.}
\usage{
mHM_reclassint(rst, lookuptbl, tif = F, asc = F, varname = "luse")
}
\arguments{
\item{rst}{raster object.}

\item{lookuptbl}{lookup table, e.g. given by \code{data()}.}

\item{tif}{boolean, if TRUE output raster written in GEOtif format.}

\item{asc}{boolean, if TRUE output raster written in ASCII format.}

\item{varname}{character, variable name.}
}
\value{
reclassified raster object.
}
\description{
\code{mHM_reclassint} reclassifies a raster map, e.g. CORINE land use, to specific classes for usage in mHM.
}
\examples{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

