% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_ts2existNC.R
\name{mHM_ts2existNC}
\alias{mHM_ts2existNC}
\title{impute timeseries in netCDF (at site).}
\usage{
mHM_ts2existNC(nc_file, var.old = "pre", var.new = "et", ts)
}
\arguments{
\item{nc_file}{netCDF file.}

\item{var.old}{character, varible name of input netCDF file, default = "pre".}

\item{var.new}{character, variable name of overwritten netCDF file, default = "et".}

\item{ts}{zoo object, time series data}
}
\value{
netCDF file, new time series data imputed.
}
\description{
\code{mHM_ts2existNC} takes an existing netCDF file at a single site and feeds it with new time series data.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
