% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_ts2netCDF.R
\name{mHM_ts2netCDF}
\alias{mHM_ts2netCDF}
\title{point calib/valid ts data to netCDF.}
\usage{
mHM_ts2netCDF(dem = "./input/morph/dem.asc",
  latlon = "./input/latlon/latlon.nc", ts, outfile, var.name, unit)
}
\arguments{
\item{dem}{morphological dummy file of mHM simulation, default: "./input/morph/dem.asc".}

\item{latlon}{latlon.nc file of mHM simulation, default: "./input/latlon/latlon.nc".}

\item{ts}{time series zoo data}

\item{outfile}{character, output file name, default: "out.nc".}

\item{var.name}{character, variable name to be set in netCDF.}

\item{unit}{character, variable unit to be set in netCDF.}
}
\value{
netCDF written to outfile.
}
\description{
\code{mHM_ts2netCDF} is converting time series data to netCDF
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
