% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_vargetmulti.R
\name{mHM_vargetmulti}
\alias{mHM_vargetmulti}
\title{get multiple variables from netCDF.}
\usage{
mHM_vargetmulti(con, vars)
}
\arguments{
\item{con}{connection to netCDF.}

\item{vars}{variable names.}
}
\value{
list containing arrays of variables.
}
\description{
\code{mHM_vargetmulti} is retrieving multiple variables from a netCDF connection.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
