% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_writeHeader.R
\name{mHM_writeHeader}
\alias{mHM_writeHeader}
\title{Write mHM header file.}
\usage{
mHM_writeHeader(rst, L2_res = 10000, res = 10000, na = -9999,
  crs_espg = 23030, writeHeader = TRUE, pre = "headerAtLevel11",
  make.griddes = FALSE, remap_py_file)
}
\arguments{
\item{rst}{raster object, which will deliver extent, e.g. DEM.}

\item{L2_res}{resolution of level L2 data, meteorological input.}

\item{res}{resolution in header file.}

\item{na}{na. value}

\item{crs_espg}{interger, projection information for \code{rst} in ESPG code.}

\item{writeHeader}{boolean, if TRUE header will be written to working directory.}

\item{pre}{character, added to header file name.}

\item{make.griddes}{boolean, if TRUE calculate .griddes info}

\item{remap_py_file}{full path python-bash script remap-grid}
}
\value{
reclassified raster object, wrote to ASCII if writeHeader is TRUE.
If make.griddes is TRUE .griddes file, wrote to working directory.
}
\description{
\code{mHM_writeHeader} writes mHM header files, needed in addition to meteo netCDF input files. Moreover,
a .griddes file can be calculated for later use with CDO to remap the netCDF meteo input according to morphological input maps.
}
\details{
For remaping the netCDF meteo input see \code{cdo --help remap}. 
The python bash script remap-grid is located in the library directory, folder inst/bash of the package \code{mHMr} (\code{path.package("mHMr")}). 
For further information type \code{./remap-grid --help} after changing directory.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
