% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr2vec.R
\name{chr2vec}
\alias{chr2vec}
\title{Helper for mHM_read.nml}
\usage{
chr2vec(chr)
}
\arguments{
\item{chr}{vector of characters to be splited and trimed.}
}
\value{
splited, trimed characters in list.
}
\description{
Helper for mHM_read.nml
}
\examples{
chr2vec()

}

