% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_avgarray2zoo.R
\name{mHM_avgarray2zoo}
\alias{mHM_avgarray2zoo}
\title{get summarized zoo object for mHM basin.}
\usage{
mHM_avgarray2zoo(array, datetime, fun, na.rm = FALSE)
}
\arguments{
\item{array}{data array.}

\item{datetime}{Date, POSIX or Yearmon object.}

\item{fun}{function to summarize/aggregate basin.}

\item{na.rm}{boolean, remove NA}
}
\value{
Zoo object.
}
\description{
\code{mHM_avgarray2zoo} is retrieving a basin summary, depicted by function \code{fun}, and giving back a zoo object.
}
\examples{

}
\references{

}
\seealso{
\code{\link{mHM_getDateTime}}
}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
