% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_enlargeNETCDF.R
\name{mHM_enlargeNETCDF}
\alias{mHM_enlargeNETCDF}
\title{enlarge NetCDF files}
\usage{
mHM_enlargeNETCDF(netcdf, x, y, var, fixcorner = 1)
}
\arguments{
\item{netcdf}{netCDF file}

\item{x}{integer, x dimension length after enlargement of field}

\item{y}{integer, y dimension length after enlargement of field}

\item{var}{name of varable to process}

\item{fixcorner}{field corner not to be moved, Default: 1. See detail}
}
\value{
NULL, new netCDF file will be written beside exiating in-netCDF
}
\description{
enlarge netCDF file in its x and y dimension
}
\details{
fixcorner 1: bottomleft, fixcorner 2: topleft, fixcorner 3: topright, fixcorner 4: bottomright
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RNetCDF]{open.nc}},\code{\link[RNetCDF]{var.get.nc}},\code{\link[RNetCDF]{create.nc}},\code{\link[RNetCDF]{dim.def.nc}},\code{\link[RNetCDF]{var.def.nc}},\code{\link[RNetCDF]{att.copy.nc}},\code{\link[RNetCDF]{var.put.nc}},\code{\link[RNetCDF]{close.nc}}
}
