% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_extrapolNetCDF.R
\name{mHM_extrapolNetCDF}
\alias{mHM_extrapolNetCDF}
\title{Simple extrapolation of netCDF in invalid locations.}
\usage{
mHM_extrapolNetCDF(netCDF, var, ext, return = FALSE)
}
\arguments{
\item{netCDF}{netCDF file name, full path.}

\item{var}{character, name of variable to extrapolate.}

\item{ext}{integer, radius (number of grid cells) within grid cells are extended, starting from border grid cells.}

\item{return}{boolean, should extrapolated array be returned in console? default=FALSE}
}
\value{
netCDF file
}
\description{
\code{mHM_extrapolNetCDF} uses mHM_extrapolMat to extrapolate its matrix values in invalid location, e.g. originating from 
meteorological input fields for mHM (netCDF format).
}
\details{
Runing mHM the first time for some netCDF products (e.g. Spain02, E-OBS) the warning
no meteorological input available for whole simulation domain will be printed. 
This function extrapolates grids in invalid locations within a radius (number of grid cells) given by \code{ext}.
Extrapolation is performed by a mean value of the surrounding cells of the invalid location.

Caution: netCDF input file will be overwritten, backup original file before use.
}
\examples{
# read netcdf file minimum air temperature for Júcar domain
 pkg_path <- path.package("mHMr")
 tmin_netcdf <- file.path(pkg_path, "data", "tmin.nc")

# open connection to netCDF
 tmin_op <-  RNetCDF::open.nc(tmin_netcdf)
# retrieve min. air temp  
 tmin <- RNetCDF::var.get.nc(tmin_op,"tmin")
# visualise original field
 image(tmin[,,1], zlim = c(2.5,4.5))

# extrapolate (radius: 1 grid cell)
 tmin_ext <- mHM_tmin_netcdf(netCDF = tmin_netcdf, vat = "tmin", ext = 1)
# visualise extrapolated field
 image(tmin_ext[,,1], zlim = c(2.5,4.5))


}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
