% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_getDateTime.R
\name{mHM_getDateTime}
\alias{mHM_getDateTime}
\title{get date(time) object from netCDF.}
\usage{
mHM_getDateTime(nc_file, start_datetime = "1950-01-01 00:00:00",
  in_unit = "h", out_unit = "d")
}
\arguments{
\item{start_datetime}{character, start datetime of netCDF, default: '1950-01-01 00:00:00'.}

\item{in_unit}{definition of datetime unit in; 'h'=hour ,'d'=day, 'm'=month.}

\item{out_unit}{definition of datetime unit out; 'h'=hour ,'d'=day, 'm'=month.}

\item{con}{connection to netCDF file, see \code{\link{open.nc}}.}
}
\value{
Chron datetime, Date or Yearmon object, depending on mHM output definition.
}
\description{
\code{mHM_getDateTime} is retrieving a date(time) object (e.g. \code{\link{chron}}, \code{\link{as.Date}}, \code{\link{as.yearmon}}) from netCDF.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
