% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_readQ.R
\name{mHM_readQ}
\alias{mHM_readQ}
\title{Read mHM daily_discharge.out in zoo object.}
\usage{
mHM_readQ(outQpath, dischargeFile = "daily_discharge.out")
}
\arguments{
\item{outQpath}{path to folder containing the file daily_discharge.}

\item{dischargeFile}{name of discharge file, default = "daily_discharge.out"}
}
\value{
daily zoo object including observed and simulated discharge.
}
\description{
Read mHM daily_discharge.out in zoo object.
}
\examples{
outQ <- mHM_readQ(outQpath="your/path")

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
