% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_reclassint.R
\name{mHM_reclassint}
\alias{mHM_reclassint}
\title{Reclassify raster with lookuptable.}
\usage{
mHM_reclassint(rst, lookuptbl, tif = F, asc = F, varname = "luse")
}
\arguments{
\item{rst}{raster object.}

\item{lookuptbl}{lookup table, e.g. given by \code{data(luse_reclass)}.}

\item{tif}{boolean, if TRUE output raster written in GEOtif format.}

\item{asc}{boolean, if TRUE output raster written in ASCII format.}

\item{varname}{character, variable name.}
}
\value{
reclassified raster object.
}
\description{
\code{mHM_reclassint} reclassifies a raster map, e.g. CORINE land use, to specific classes for usage in mHM.
}
\examples{
library(raster)

data(corine_jucar)
corine_jucar <- as.raster(corine_jucar)
plot(corine_jucar)

data(luse_reclass)
luse_reclass <- as.matrix(luse_reclass)

mHMluse_jucar <- mHM_reclassint(rst=corine_jucar, lookuptbl=luse_reclass_tbl, tif=F, asc=F, varname="luse")
plot(mHMluse_jucar)

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\references{

}
\seealso{

}
\keyword{}

