% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHM_ts2existNC.R
\name{mHM_ts2existNC}
\alias{mHM_ts2existNC}
\title{impute timeseries in netCDF (at site).}
\usage{
mHM_ts2existNC(nc_file, var, ts)
}
\arguments{
\item{nc_file}{connection to netCDF file, see \code{\link{open.nc}}.}

\item{ts}{zoo object, time series data}
}
\value{
netCDF file, new time series data imputed.
}
\description{
\code{mHM_ts2existNC} takes an existing netCDF file at a single site and feeds it with new time series data.
}
\examples{

}
\references{

}
\seealso{

}
\author{
Johannes Brenner \email{johannes.brenner@ufz.de}
}
\keyword{}
